pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docx2html"
version = "0.1.0"
description = "Convert Microsoft Word (.docx) files to clean, structured HTML optimised for LLM ingestion."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "docx2html contributors" }]
keywords = ["docx", "html", "llm", "converter", "word"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: HTML",
    "Topic :: Utilities",
]

dependencies = [
    "python-docx>=1.1.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "python-docx>=1.1.0",
]

[project.scripts]
docx2html = "docx2html.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["docx2html*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["docx2html"]
omit = ["tests/*"]
